requirements.txt
# Conscious Child AI - Python Dependencies

# ============================================================================
# Core Framework
# ============================================================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ============================================================================
# Deep Learning & AI
# ============================================================================
torch==2.1.0
transformers==4.35.2
sentence-transformers==2.2.2
accelerate==0.25.0
bitsandbytes==0.41.0

# LLM Inference
llama-cpp-python==0.2.20  # For local LLM
vllm==0.2.6  # Optional: faster inference

# ============================================================================
# Speech Processing
# ============================================================================
openai-whisper==20231117
faster-whisper==0.10.0  # Optimized Whisper
TTS==0.20.0  # Coqui TTS

# Audio processing
soundfile==0.12.1
librosa==0.10.1
pydub==0.25.1

# ============================================================================
# Database & Storage
# ============================================================================
psycopg2-binary==2.9.9
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.13.0  # Database migrations

redis==5.0.1
hiredis==2.2.3  # Fast Redis parser

chromadb==0.4.18  # Vector database

# ============================================================================
# Internet & Web Access
# ============================================================================
httpx==0.25.2
requests==2.31.0
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3
readability-lxml==0.8.1

# Search
duckduckgo-search==3.9.6

# ============================================================================
# Security & Authentication
# ============================================================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7

# ============================================================================
# Monitoring & Logging
# ============================================================================
structlog==23.2.0
python-json-logger==2.0.7
prometheus-client==0.19.0

# ============================================================================
# Utilities
# ============================================================================
python-dotenv==1.0.0
pyyaml==6.0.1
toml==0.10.2
pendulum==2.1.2  # Better datetime
python-dateutil==2.8.2

# Task scheduling
apscheduler==3.10.4

# Progress bars
tqdm==4.66.1

# Data validation
email-validator==2.1.0

# ============================================================================
# Development & Testing
# ============================================================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.0
ruff==0.1.8
mypy==1.7.1

# ============================================================================
# Symbolic Reasoning
# ============================================================================
pyDatalog==0.17.1  # Prolog-like reasoning
networkx==3.2.1  # Knowledge graphs

# ============================================================================
# Reinforcement Learning (for value learning)
# ============================================================================
gymnasium==0.29.1

# ============================================================================
# Misc
# ============================================================================
numpy==1.26.2
scipy==1.11.4
pandas==2.1.4
scikit-learn==1.3.2

